refactor: collapse health/ready service modules into a shared macro and tighten pub visibility in peppylib-py
