+++ crates/peppylib-py/Cargo.toml
@@ -8,7 +8,6 @@ name = "_peppylib"
 crate-type = ["cdylib"]
 
 [dependencies]
-bytes = "1"
 config = { path = "../config-internal" }
 peppylib = { path = "../peppylib" }
 pmi = { path = "../pmi-internal", features = ["zenoh", "build_zenoh"] }
@@ -19,7 +18,6 @@ rand = "0.10"
 names-generator2 = "0.2.1"
 serde_json = "1"
 tokio = { version = "1.47", features = ["rt-multi-thread"] }
-tokio-util = "0.7"
 
 [build-dependencies]
 build-helpers = { path = "../build-helpers-internal" }
